Avoid code duplication by using abstract base class for tests.
